+++ src/dftbp/dftbplus/inputdata.F90
@@ -26,6 +26,7 @@ module dftbp_dftbplus_inputdata
   use dftbp_dftb_slakocont, only : TSlakoCont
   use dftbp_dftbplus_input_geoopt, only : TGeoOptInput
   use dftbp_elecsolvers_elecsolvers, only : TElectronicSolverInp
+  use dftbp_extlibs_externalmodel, only : TExtModelProvides
   use dftbp_extlibs_poisson, only : TPoissonInfo
   use dftbp_extlibs_tblite, only : TTBLiteInput
   use dftbp_io_message, only : error, warning
@@ -144,6 +145,9 @@ module dftbp_dftbplus_inputdata
 
     !> Choice of electronic hamiltonian
     integer :: hamiltonian = hamiltonianTypes%none
+
+    !> External model to the main code (if used)
+    type(TExtModelProvides), allocatable :: extModel
 
     !> Random number generator seed
     integer :: iSeed = 0