+++ src/dftbp/dftbplus/main.F90
@@ -98,6 +98,7 @@ module dftbp_dftbplus_main
   use dftbp_md_xlbomd, only : TXLBOMD
   use dftbp_mixer_mixer, only : TMixerReal, TMixerCmplx, TMixerReal_reset, TMixerCmplx_reset,&
       & TMixerReal_mix, TMixerCmplx_mix, TMixerReal_getInverseJacobian
+  use dftbp_modelindependent_localclusters, only : TLocalClusters
   use dftbp_reks_reks, only : TReksCalc, guessneweigvecs, optimizeFONs, calcweights, activeorbswap,&
       & getfilling, calcsareksenergy, printsareksenergy, qm2udl, printreksmicrostates, qmexpandl,&
       & ud2qml, constructmicrostates, checkgammapoint, getfockanddiag, printrekssainfo,&
@@ -1086,11 +1087,18 @@ contains
     !! Whether constraints are converged
     logical :: constrConverged
 
+    ! Local environments around atomic sites
+    type(TLocalClusters), allocatable, target :: localClusters
+
     integer :: iKS, iConstrIter, nConstrIter
     logical :: isFirstDet
 
     if (this%tDipole) allocate(dipoleTmp(3))
     isFirstDet = this%deltaDftb%iDeterminant == 1
+
+    if (allocated(this%localClusters)) then
+      localClusters = this%localClusters
+    end if
 
     call env%globalTimer%startTimer(globalTimers%preSccInit)
 
@@ -1125,6 +1133,11 @@ contains
         @:PROPAGATE_ERROR(errStatus)
     end if
 
+    if (allocated(localClusters) .and. (this%tLatticeChanged .or. this%tCoordsChanged)) then
+      call localClusters%updateGeometry(this%coord, this%neighbourList,&
+          & this%img2CentCell, this%species)
+    end if
+
   #:if WITH_TRANSPORT
     if (this%tNegf .and. isFirstDet) then
       call setupNegfStuff(this%negfInt, this%denseDesc, this%transpar, this%ginfo,&
@@ -1153,6 +1166,17 @@ contains
           & this%denseDesc%iAtomStart, this%iSparseStart, this%img2CentCell)
     end if
 
+    if (allocated(localClusters)) then
+      call localClusters%generateClusters(this%coord, this%nNeighbourSK, this%neighbourList,&
+          & this%img2CentCell, this%species, this%speciesName)
+    end if
+
+    if (this%hamiltonianType == hamiltonianTypes%api) then
+      call this%extModel%update(this%species, localClusters, this%iSparseStart, this%nNeighbourSK,&
+          & errStatus)
+      @:PROPAGATE_ERROR(errStatus)
+    end if
+
     if (isFirstDet) then
       call env%globalTimer%startTimer(globalTimers%sparseH0S)
       select case(this%hamiltonianType)
@@ -1169,12 +1193,14 @@ contains
             & this%neighbourList%iNeighbour, this%img2CentCell, this%iSparseStart, &
             & this%orb, this%H0, this%ints%overlap, this%ints%dipoleBra, this%ints%dipoleKet, &
             & this%ints%quadrupoleBra, this%ints%quadrupoleKet)
+      case(hamiltonianTypes%api)
+        call this%extModel%buildSH0(env, this%H0, this%ints%overlap, this%nAtom, this%species,&
+            & this%iSparseStart, this%orb, errStatus)
+        @:PROPAGATE_ERROR(errStatus)
       end select
       call env%globalTimer%stopTimer(globalTimers%sparseH0S)
 
-      if (this%tSetFillingTemp) then
-        call this%temperatureProfile%getTemperature(this%tempElec)
-      end if
+      if (this%tSetFillingTemp) call this%temperatureProfile%getTemperature(this%tempElec)
       call this%electronicSolver%updateElectronicTemp(this%tempElec)
     end if