+++ CHANGELOG.rst
@@ -108,6 +108,11 @@ Added
 
 - Dual American and British English spelling for various input keywords
 
+- External models (electronic or energy based) from linked (static or
+  dynamic) library outside the code base.
+
+- Rename "Hamiltonian" to "Model" in input, as more descriptive
+
 
 Fixed
 -----